requirements.txt
aiofiles==24.1.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.5.7
decorator==5.1.1
deprecation==2.1.0
dnspython==2.6.1
email_validator==2.2.0
eval_type_backport==0.2.0
exceptiongroup==1.2.2
executing==2.0.1
fastapi==0.111.1
fastapi-cli==0.0.4
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.6.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.1
idna==3.7
ipyplot==1.1.2
ipython==8.26.0
jedi==0.19.1
Jinja2==3.1.4
lorax-client==0.6.1
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
marshmallow-enum==1.5.1
matplotlib-inline==0.1.7
mdurl==0.1.2
motor==3.5.1
multidict==6.0.5
mypy-extensions==1.0.0
numpy==2.0.1
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
predibase==2024.7.2
predibase-api==2024.7.2
progress-table==0.1.26
prompt_toolkit==3.0.47
protobuf==5.27.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pydantic==2.8.2
pydantic-settings==2.3.4
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.8.0
pymongo==4.8.0
PyMuPDF==1.24.8
PyMuPDFb==1.24.8
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.9
python-pptx==0.6.23
python-rapidjson==1.18
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
rich==13.7.1
semantic-version==2.10.0
shellingham==1.5.4
shortuuid==1.0.13
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.37.2
tabulate==0.9.0
tqdm==4.66.4
traitlets==5.14.3
tritonclient==2.41.1
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.12
uvicorn==0.30.3
uvloop==0.19.0
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
XlsxWriter==3.2.0
yarl==1.9.4
